Update: Project structure - Added folder for services only - Update logic not on component
