+++ src/router/index.ts
@@ -1,6 +1,7 @@
 import { createRouter, createWebHistory } from "vue-router";
 // Import your Todos component
 import HomeView from "@/views/HomeView.vue";
+import UserView from "@/views/UserView.vue";
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -13,7 +14,7 @@ const router = createRouter({
     {
       path: "/detailTodos",
       name: "detail",
-      component: HomeView,
+      component: UserView,
     },
     // {
     //   path: '/user',
